requirements.txt
python-jobspy
pandas
gspread
oauth2client
openpyxl
telethon
playwright
jinja2
google-generativeai
pytest-playwright
